refactor: clean up RBAC interfaces into RBAC module
