+++ src/ai/backend/manager/models/vfolder.py
@@ -4,8 +4,7 @@ import enum
 import logging
 import os.path
 import uuid
-from collections.abc import Container, Mapping
-from dataclasses import dataclass
+from collections.abc import Mapping
 from datetime import datetime
 from pathlib import PurePosixPath
 from typing import (
@@ -16,7 +15,6 @@ from typing import (
     NamedTuple,
     Optional,
     Sequence,
-    TypeAlias,
     cast,
 )
 
@@ -86,22 +84,6 @@ from .gql_relay import AsyncNode, Connection, ConnectionResolverResult
 from .group import GroupRow, ProjectType
 from .minilang.ordering import OrderSpecItem, QueryOrderParser
 from .minilang.queryfilter import FieldSpecItem, QueryFilterParser, enum_field_getter
-from .rbac import (
-    AbstractPermissionContext,
-    AbstractPermissionContextBuilder,
-    BasePermission,
-    BaseScope,
-    ClientContext,
-    DomainScope,
-    ProjectScope,
-    ScopedUserRole,
-    StorageHost,
-    get_roles_in_scope,
-)
-from .rbac import (
-    UserScope as UserRBACScope,
-)
-from .rbac.exceptions import InvalidScope, NotEnoughPermission
 from .session import DEAD_SESSION_STATUSES, SessionRow
 from .user import UserRole, UserRow
 from .utils import ExtendedAsyncSAEngine, execute_with_retry, sql_json_merge
@@ -116,7 +98,6 @@ __all__: Sequence[str] = (
     "vfolder_invitations",
     "vfolder_permissions",
     "VirtualFolder",
-    "VFolderRBACPermission",
     "VFolderOwnershipType",
     "VFolderInvitationState",
     "VFolderPermission",
@@ -145,11 +126,6 @@ __all__: Sequence[str] = (
     "SOFT_DELETED_VFOLDER_STATUSES",
     "HARD_DELETED_VFOLDER_STATUSES",
     "VFolderPermissionSetAlias",
-    "get_vfolders",
-    "VFolderWithPermissionSet",
-    "OWNER_PERMISSIONS",
-    "PermissionContext",
-    "PermissionContextBuilder",
 )
 
 
@@ -702,429 +678,6 @@ async def query_accessible_vfolders(
                 entry["permission"] = overriding_permissions[entry["id"]]
 
     return entries
-
-
-class VFolderRBACPermission(BasePermission):
-    # Only owners can do
-    CLONE = enum.auto()
-    ASSIGN_PERMISSION_TO_OTHERS = enum.auto()  # Invite, share
-
-    # `create_vfolder` action should be in {Domain, Project, or User} permissions, not here
-    READ_ATTRIBUTE = enum.auto()
-    UPDATE_ATTRIBUTE = enum.auto()
-    DELETE_VFOLDER = enum.auto()
-
-    READ_CONTENT = enum.auto()
-    WRITE_CONTENT = enum.auto()
-    DELETE_CONTENT = enum.auto()
-
-    MOUNT_RO = enum.auto()
-    MOUNT_RW = enum.auto()
-    MOUNT_WD = enum.auto()
-
-
-WhereClauseType: TypeAlias = (
-    sa.sql.expression.BinaryExpression | sa.sql.expression.BooleanClauseList
-)
-# TypeAlias is deprecated since 3.12
-
-OWNER_PERMISSIONS: frozenset[VFolderRBACPermission] = frozenset([
-    perm for perm in VFolderRBACPermission
-])
-ADMIN_PERMISSIONS: frozenset[VFolderRBACPermission] = frozenset([
-    VFolderRBACPermission.READ_ATTRIBUTE,
-    VFolderRBACPermission.UPDATE_ATTRIBUTE,
-    VFolderRBACPermission.DELETE_VFOLDER,
-])
-MONITOR_PERMISSIONS: frozenset[VFolderRBACPermission] = frozenset([
-    VFolderRBACPermission.READ_ATTRIBUTE,
-    VFolderRBACPermission.UPDATE_ATTRIBUTE,
-])
-PRIVILEGED_MEMBER_PERMISSIONS: frozenset[VFolderRBACPermission] = frozenset([
-    VFolderRBACPermission.READ_ATTRIBUTE,
-    VFolderRBACPermission.READ_CONTENT,
-    VFolderRBACPermission.WRITE_CONTENT,
-    VFolderRBACPermission.DELETE_CONTENT,
-    VFolderRBACPermission.MOUNT_RO,
-    VFolderRBACPermission.MOUNT_RW,
-    VFolderRBACPermission.MOUNT_WD,
-])
-MEMBER_PERMISSIONS: frozenset[VFolderRBACPermission] = frozenset()
-
-# TODO: Change type of `vfolder_permissions.permission` to VFolderRBACPermission
-PERMISSION_TO_RBAC_PERMISSION_MAP: Mapping[VFolderPermission, frozenset[VFolderRBACPermission]] = {
-    VFolderPermission.READ_ONLY: frozenset([
-        VFolderRBACPermission.READ_ATTRIBUTE,
-        VFolderRBACPermission.READ_CONTENT,
-    ]),
-    VFolderPermission.READ_WRITE: frozenset([
-        VFolderRBACPermission.READ_ATTRIBUTE,
-        VFolderRBACPermission.UPDATE_ATTRIBUTE,
-        VFolderRBACPermission.DELETE_VFOLDER,
-        VFolderRBACPermission.READ_CONTENT,
-        VFolderRBACPermission.WRITE_CONTENT,
-        VFolderRBACPermission.DELETE_CONTENT,
-        VFolderRBACPermission.MOUNT_RO,
-        VFolderRBACPermission.MOUNT_RW,
-    ]),
-    VFolderPermission.RW_DELETE: frozenset([
-        VFolderRBACPermission.READ_ATTRIBUTE,
-        VFolderRBACPermission.UPDATE_ATTRIBUTE,
-        VFolderRBACPermission.DELETE_VFOLDER,
-        VFolderRBACPermission.READ_CONTENT,
-        VFolderRBACPermission.WRITE_CONTENT,
-        VFolderRBACPermission.DELETE_CONTENT,
-        VFolderRBACPermission.MOUNT_RO,
-        VFolderRBACPermission.MOUNT_RW,
-        VFolderRBACPermission.MOUNT_WD,
-    ]),
-    VFolderPermission.OWNER_PERM: frozenset(OWNER_PERMISSIONS),
-}
-
-
-@dataclass
-class PermissionContext(AbstractPermissionContext[VFolderRBACPermission, VFolderRow, uuid.UUID]):
-    @property
-    def query_condition(self) -> WhereClauseType | None:
-        cond: WhereClauseType | None = None
-
-        def _OR_coalesce(
-            base_cond: WhereClauseType | None,
-            _cond: sa.sql.expression.BinaryExpression,
-        ) -> WhereClauseType:
-            return base_cond | _cond if base_cond is not None else _cond
-
-        def _AND_coalesce(
-            base_cond: WhereClauseType | None,
-            _cond: sa.sql.expression.BinaryExpression,
-        ) -> WhereClauseType:
-            return base_cond & _cond if base_cond is not None else _cond
-
-        if self.user_id_to_permission_map:
-            cond = _OR_coalesce(cond, VFolderRow.user.in_(self.user_id_to_permission_map.keys()))
-        if self.project_id_to_permission_map:
-            cond = _OR_coalesce(
-                cond, VFolderRow.group.in_(self.project_id_to_permission_map.keys())
-            )
-        if self.domain_name_to_permission_map:
-            cond = _OR_coalesce(
-                cond, VFolderRow.domain_name.in_(self.domain_name_to_permission_map.keys())
-            )
-        if self.object_id_to_additional_permission_map:
-            cond = _OR_coalesce(
-                cond, VFolderRow.id.in_(self.object_id_to_additional_permission_map.keys())
-            )
-        if self.object_id_to_overriding_permission_map:
-            cond = _OR_coalesce(
-                cond, VFolderRow.id.in_(self.object_id_to_overriding_permission_map.keys())
-            )
-
-        return cond
-
-    async def build_query(self) -> sa.sql.Select | None:
-        cond = self.query_condition
-        if cond is None:
-            return None
-        return sa.select(VFolderRow).where(cond)
-
-    async def calculate_final_permission(
-        self, rbac_obj: VFolderRow
-    ) -> frozenset[VFolderRBACPermission]:
-        vfolder_row = rbac_obj
-        vfolder_id = cast(uuid.UUID, vfolder_row.id)
-        if (
-            overriding_perm := self.object_id_to_overriding_permission_map.get(vfolder_id)
-        ) is not None:
-            return overriding_perm
-        permissions: set[VFolderRBACPermission] = set()
-        permissions |= self.object_id_to_additional_permission_map.get(vfolder_id, set())
-        permissions |= self.user_id_to_permission_map.get(vfolder_row.user, set())
-        permissions |= self.project_id_to_permission_map.get(vfolder_row.group, set())
-        permissions |= self.domain_name_to_permission_map.get(vfolder_row.domain_name, set())
-        return frozenset(permissions)
-
-
-class PermissionContextBuilder(
-    AbstractPermissionContextBuilder[VFolderRBACPermission, PermissionContext]
-):
-    db_session: SASession
-
-    def __init__(self, db_session: SASession) -> None:
-        self.db_session = db_session
-
-    async def build(
-        self,
-        ctx: ClientContext,
-        target_scope: BaseScope,
-        requested_permission: VFolderRBACPermission,
-    ) -> PermissionContext:
-        match target_scope:
-            case DomainScope(domain_name):
-                permission_ctx = await self.build_in_domain_scope(ctx, domain_name)
-            case ProjectScope(project_id, domain_name):
-                permission_ctx = await self.build_in_project_scope(ctx, project_id)
-            case UserRBACScope(user_id, _):
-                permission_ctx = await self.build_in_user_scope(ctx, user_id)
-            case _:
-                raise InvalidScope
-        permission_ctx.filter_by_permission(requested_permission)
-        return permission_ctx
-
-    async def build_in_nested_scope(
-        self,
-        ctx: ClientContext,
-        target_scope: BaseScope,
-        requested_permission: VFolderRBACPermission,
-    ) -> PermissionContext:
-        match target_scope:
-            case DomainScope(domain_name):
-                permission_ctx = await self.build_in_domain_scope(ctx, domain_name)
-                _user_perm_ctx = await self.build_in_user_scope_in_domain(
-                    ctx, ctx.user_id, domain_name
-                )
-                permission_ctx = PermissionContext.merge(permission_ctx, _user_perm_ctx)
-                _project_perm_ctx = await self.build_in_project_scopes_in_domain(ctx, domain_name)
-                permission_ctx = PermissionContext.merge(permission_ctx, _project_perm_ctx)
-            case ProjectScope(project_id, domain_name):
-                permission_ctx = await self.build_in_project_scope(ctx, project_id)
-                if domain_name is not None:
-                    _user_perm_ctx = await self.build_in_user_scope_in_domain(
-                        ctx, ctx.user_id, domain_name
-                    )
-                    permission_ctx = PermissionContext.merge(permission_ctx, _user_perm_ctx)
-            case UserRBACScope(user_id, _):
-                permission_ctx = await self.build_in_user_scope(ctx, user_id)
-            case _:
-                raise InvalidScope
-        permission_ctx.filter_by_permission(requested_permission)
-        return permission_ctx
-
-    async def build_in_domain_scope(
-        self,
-        ctx: ClientContext,
-        domain_name: str,
-    ) -> PermissionContext:
-        roles = await get_roles_in_scope(ctx, DomainScope(domain_name), self.db_session)
-        domain_permissions = await PermissionContextBuilder.calculate_permission_by_roles(roles)
-        result = PermissionContext(domain_name_to_permission_map={domain_name: domain_permissions})
-        return result
-
-    async def build_in_project_scopes_in_domain(
-        self,
-        ctx: ClientContext,
-        domain_name: str,
-    ) -> PermissionContext:
-        result = PermissionContext()
-
-        _project_stmt = (
-            sa.select(GroupRow)
-            .where(GroupRow.domain_name == domain_name)
-            .options(load_only(GroupRow.id))
-        )
-        for row in await self.db_session.scalars(_project_stmt):
-            _row = cast(GroupRow, row)
-            _project_perm_ctx = await self.build_in_project_scope(ctx, _row.id)
-            result = PermissionContext.merge(result, _project_perm_ctx)
-        return result
-
-    async def build_in_user_scope_in_domain(
-        self,
-        ctx: ClientContext,
-        user_id: uuid.UUID,
-        domain_name: str,
-    ) -> PermissionContext:
-        # For Superadmin and monitor who can create vfolders in multiple different domains.
-        roles = await get_roles_in_scope(ctx, UserRBACScope(user_id, domain_name), self.db_session)
-        permissions = await PermissionContextBuilder.calculate_permission_by_roles(roles)
-
-        _vfolder_stmt = (
-            sa.select(VFolderRow)
-            .where((VFolderRow.user == user_id) & (VFolderRow.domain_name == domain_name))
-            .options(load_only(VFolderRow.id))
-        )
-        own_folder_map = {
-            row.id: permissions for row in await self.db_session.scalars(_vfolder_stmt)
-        }
-        result = PermissionContext(object_id_to_additional_permission_map=own_folder_map)
-
-        _stmt = (
-            sa.select(VFolderPermissionRow)
-            .select_from(sa.join(VFolderPermissionRow, VFolderRow))
-            .where(
-                (VFolderPermissionRow.user == ctx.user_id)
-                & (
-                    VFolderRow.ownership_type == VFolderOwnershipType.USER
-                )  # filter out user vfolders
-                & (VFolderRow.domain_name == domain_name)
-            )
-        )
-        object_id_to_permission_map = {
-            row.vfolder: PERMISSION_TO_RBAC_PERMISSION_MAP[row.permission]
-            for row in await self.db_session.scalars(_stmt)
-        }
-        if ctx.user_role in (UserRole.SUPERADMIN, UserRole.ADMIN):
-            ctx_to_merge = PermissionContext(
-                object_id_to_additional_permission_map=object_id_to_permission_map
-            )
-        else:
-            ctx_to_merge = PermissionContext(
-                object_id_to_overriding_permission_map=object_id_to_permission_map
-            )
-        result = PermissionContext.merge(result, ctx_to_merge)
-        return result
-
-    async def build_in_project_scope(
-        self,
-        ctx: ClientContext,
-        project_id: uuid.UUID,
-    ) -> PermissionContext:
-        roles = await get_roles_in_scope(ctx, ProjectScope(project_id), self.db_session)
-        permissions = await PermissionContextBuilder.calculate_permission_by_roles(roles)
-        result = PermissionContext(project_id_to_permission_map={project_id: permissions})
-
-        _stmt = (
-            sa.select(VFolderPermissionRow)
-            .select_from(sa.join(VFolderPermissionRow, VFolderRow))
-            .where(
-                (VFolderPermissionRow.user == ctx.user_id)
-                & (
-                    VFolderRow.ownership_type == VFolderOwnershipType.GROUP
-                )  # filter out user vfolders
-            )
-        )
-        object_id_to_permission_map = {
-            row.vfolder: PERMISSION_TO_RBAC_PERMISSION_MAP[row.permission]
-            for row in await self.db_session.scalars(_stmt)
-        }
-        if ScopedUserRole.ADMIN in roles:
-            result.object_id_to_additional_permission_map = object_id_to_permission_map
-        else:
-            result.object_id_to_overriding_permission_map = object_id_to_permission_map
-        return result
-
-    async def build_in_user_scope(
-        self,
-        ctx: ClientContext,
-        user_id: uuid.UUID,
-    ) -> PermissionContext:
-        roles = await get_roles_in_scope(ctx, UserRBACScope(user_id), self.db_session)
-        permissions = await PermissionContextBuilder.calculate_permission_by_roles(roles)
-        result = PermissionContext(user_id_to_permission_map={user_id: permissions})
-
-        _stmt = (
-            sa.select(VFolderPermissionRow)
-            .select_from(sa.join(VFolderPermissionRow, VFolderRow))
-            .where(
-                (VFolderPermissionRow.user == ctx.user_id)
-                & (
-                    VFolderRow.ownership_type == VFolderOwnershipType.USER
-                )  # filter out user vfolders
-            )
-        )
-        object_id_to_permission_map = {
-            row.vfolder: PERMISSION_TO_RBAC_PERMISSION_MAP[row.permission]
-            for row in await self.db_session.scalars(_stmt)
-        }
-        if ctx.user_role in (UserRole.SUPERADMIN, UserRole.ADMIN):
-            result.object_id_to_additional_permission_map = object_id_to_permission_map
-        else:
-            result.object_id_to_overriding_permission_map = object_id_to_permission_map
-        return result
-
-    @classmethod
-    async def _permission_for_owner(
-        cls,
-    ) -> frozenset[VFolderRBACPermission]:
-        return OWNER_PERMISSIONS
-
-    @classmethod
-    async def _permission_for_admin(
-        cls,
-    ) -> frozenset[VFolderRBACPermission]:
-        return ADMIN_PERMISSIONS
-
-    @classmethod
-    async def _permission_for_monitor(
-        cls,
-    ) -> frozenset[VFolderRBACPermission]:
-        return MONITOR_PERMISSIONS
-
-    @classmethod
-    async def _permission_for_privileged_member(
-        cls,
-    ) -> frozenset[VFolderRBACPermission]:
-        return PRIVILEGED_MEMBER_PERMISSIONS
-
-    @classmethod
-    async def _permission_for_member(
-        cls,
-    ) -> frozenset[VFolderRBACPermission]:
-        return MEMBER_PERMISSIONS
-
-
-class VFolderWithPermissionSet(NamedTuple):
-    vfolder_row: VFolderRow
-    permissions: frozenset[VFolderRBACPermission]
-
-
-async def get_vfolders(
-    db_conn: SAConnection,
-    ctx: ClientContext,
-    target_scope: BaseScope,
-    requested_permission: VFolderRBACPermission,
-    extra_scope: StorageHost | None = None,
-    *,
-    vfolder_id: uuid.UUID | None = None,
-    vfolder_name: str | None = None,
-    usage_mode: VFolderUsageMode | None = None,
-    allowed_status: Container[VFolderOperationStatus] | None = None,
-    blocked_status: Container[VFolderOperationStatus] | None = None,
-) -> list[VFolderWithPermissionSet]:
-    async with ctx.db.begin_readonly_session(db_conn) as db_session:
-        builder = PermissionContextBuilder(db_session)
-        permission_ctx = await builder.build(ctx, target_scope, requested_permission)
-
-        query_stmt = await permission_ctx.build_query()
-        if query_stmt is None:
-            return []
-        if vfolder_id is not None:
-            query_stmt = query_stmt.where(VFolderRow.id == vfolder_id)
-        if vfolder_name is not None:
-            query_stmt = query_stmt.where(VFolderRow.name == vfolder_name)
-        if usage_mode is not None:
-            query_stmt = query_stmt.where(VFolderRow.usage_mode == usage_mode)
-        if allowed_status is not None:
-            query_stmt = query_stmt.where(VFolderRow.status.in_(allowed_status))
-        if blocked_status is not None:
-            query_stmt = query_stmt.where(VFolderRow.status.not_in(blocked_status))
-
-        result: list[VFolderWithPermissionSet] = []
-        for row in await db_session.scalars(query_stmt):
-            row = cast(VFolderRow, row)
-            permissions = await permission_ctx.calculate_final_permission(row)
-            result.append(VFolderWithPermissionSet(row, permissions))
-        return result
-
-
-async def validate_permission(
-    db_conn: SAConnection,
-    ctx: ClientContext,
-    target_scope: BaseScope,
-    extra_scope: StorageHost | None = None,
-    *,
-    permission: VFolderRBACPermission,
-    vfolder_id: uuid.UUID,
-) -> None:
-    vfolders = await get_vfolders(
-        db_conn,
-        ctx,
-        target_scope,
-        permission,
-        extra_scope,
-        vfolder_id=vfolder_id,
-    )
-    if not vfolders:
-        raise NotEnoughPermission(f"'{permission.name}' not allowed in {str(target_scope)}")
 
 
 async def get_allowed_vfolder_hosts_by_group(
